\name{ass7pack-package}
\alias{ass7pack-package}
\alias{ass7pack}
\docType{package}
\title{
Package for ERIM Programming Course Assignment 7

}
\description{
The package contains 4 superficial functions and 1 substantive function. The superficial function are there to illustrate and practice the inclusion of multiple function in a package. The substantive function can be used to count factors in a vector and return this as a matrix. This function can be applied to the data set that is accompanied with this package.
}
\details{
\tabular{ll}{
Package: \tab ass7pack\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-12-15\cr
License: \tab GLP (=< 2)\cr
}
}
\section{Contents}{
\tabular{ll}{
  Haha \tab The Haha function for math jokes \cr
  AddTwo \tab Function for adding 2 to supplied number \cr
  AddFour \tab Function for adding 4 to supplied number \cr
  Count_Factor \tab Function for counting factors in a vector and returning them as a matrix \cr
  is.funny \tab Function to assess whether something is funny \cr
  
  projectmanagers \tab Data consisting of names of project managers and technology managers over 1263 projects.
}
}

\section{Author}{
T.H. van Balen (374663tb@eur.nl)
}

