\name{Haha}
\alias{Haha}

\title{
Haha
}
\description{
The Haha function returns very funny math jokes based on the number (n) that is supplied.
}
\usage{
Haha(n)
}
\arguments{
  \item{n}{
is a real number higher than 0.
}
}
\details{
Make sure that n > 0. The length of n = 1 and that n is strictly numeric. 
Most of all a sense of humor is required to run this function.
}
\value{
The function returns very funny math jokes based on the supplied n.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
T.H. van Balen (374663tb@eur.nl)
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
Haha(1)
Haha(2)
Haha(3)
Haha(7)
Haha(8)
Haha(10)
Haha(11)
Haha(50)
Haha(1001)
Haha(304003230)
}

