\name{is.funny}
\alias{is.funny}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
is.funny
}
\description{
Function returns the logical value TRUE for things that are funny.
}
\usage{
is.funny(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
is something that might be funny.
}
}
\details{
Make sure that the what is supplied is really funny because the function is not really accurate. Use of this function is at your own risk.
}
\value{
Actually the function always returns the value TRUE. It's all a lie.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Nobody! I'm not responsible for this failure of a function.
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
x <- "What do you call counterfeited German currency? Question marks."

is.funny(x)
  
  
y <- "What did one computer say to the other? 010101101010101010101"

is.funny(y)
}